requirements.txt
boto3==1.15.6
botocore==1.18.6
Django==2.2.16
django-filter==2.3.0
django-storages==1.10.1
djangorestframework==3.11.1
gunicorn==20.0.4
jmespath==0.10.0
Markdown==3.2.2
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
